move tests to token package
